requirements.txt
fastapi
uvicorn[standard]
httpx
requests
pytest
python-dotenv
slowapi
cachetools
expiring-dict
fastapi-login
python-multipart
nose